+++ lib/Conversion/CMakeLists.txt
@@ -1,2 +1,1 @@
 add_subdirectory(PTensorToLinalg)
-add_subdirectory(DistElim)
